refactor(ode): tableau-driven explicit-RK family; Tsit5 measured and declined
Generalize the bespoke Vern7 stepper into `ErkTableau` + `ErkStepper`, so an
explicit Runge-Kutta method is a `const` rather than an implementation.
Vern7 is unchanged behaviourally (it now reuses `ks[0]` across a rejected
step, which saves an evaluation without changing any value).

Tsitouras 5(4) was the motivation: same 7 stages and FSAL structure as the
default Dormand-Prince tableau, smaller truncation-error constant, so it
looked like a free upgrade. Measured against the Savic transit anchor at
1e-9 it needed 3720 accepted steps against RK45's 3940 — ~6% — and was
slower in wall-clock, because `Rk45Stepper` is hand-unrolled while the
generic stepper walks index loops. Vern7's 64% step reduction clears that
overhead; Tsit5's 6% does not, so it is not exposed. The measurement and the
route to reviving it (express RK45 itself as an `ErkTableau` and retire the
unrolled copy, accepting a re-anchoring of the default path) are recorded in
the module docs so the work is not silently redone.
