+++ src/ode/mod.rs
@@ -1,8 +1,8 @@
 pub(crate) mod ekf;
+pub(crate) mod explicit_rk;
 pub mod predictions;
 pub(crate) mod rosenbrock;
 pub mod solver;
-pub(crate) mod verner;
 
 pub use predictions::*;
 pub use solver::*;
